% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_avg_Z_score.R
\name{calculate_avg_z_score}
\alias{calculate_avg_z_score}
\title{Calculate Average Z Score
Calculates the average z score for any given genes within a gene score}
\usage{
calculate_avg_z_score(gene_matrix, gene_list)
}
\arguments{
\item{gene_matrix}{a data frame where columns are Gene Symbols and specific genes}

\item{gene_list}{a list of genes within a certain gene symbol}
}
\description{
Calculate Average Z Score
Calculates the average z score for any given genes within a gene score
}
