% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateBuffa.R
\name{calculateBuffa}
\alias{calculateBuffa}
\title{Calculate the Buffa hypoxia score}
\usage{
calculateBuffa(genemat, buffa.genes = inputGenes("Buffa"))
}
\arguments{
\item{genemat}{A data frame of expression counts where rows are genes, columns are samples, and the gene symbols are available in a column defined as "Gene".}

\item{buffa.genes}{A character vector of the genes to use when calculating the score. May need to be altered due to changing gene labels.}
}
\value{
A data frame where a column of samples are reported with a column of calcuated Buffa scores for the set.
}
\description{
This is a score determine by the ranked expressions of 51 genes associated with hypoxia. The final
scores are dependent on the provided sample set, not uniquely calculated for each sample.
}
