% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMitoscore.R
\name{calculateMitoscore}
\alias{calculateMitoscore}
\title{Calculate the Mitoscore}
\usage{
calculateMitoscore(genemat, mito.genes = inputGenes("Mitoscore"))
}
\arguments{
\item{genemat}{A data frame of expression counts where rows are genes, columns are samples, and the gene symbols are available in a column defined as "Gene".}

\item{mito.genes}{A character vector of the genes to use when calculating the score. May need to be altered due to changing gene labels.}
}
\value{
A data frame with a column of samples and a column of their associated mitoscores.
}
\description{
The mitoscore is a measure of the average expression of the logged counts of a number of genes associated with hypoxia.
}
