% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateOCRscore.R
\name{calculateOCRscore}
\alias{calculateOCRscore}
\title{Calculate the OCR-score}
\usage{
calculateOCRscore(gene_matrix, ocr.genes = inputGenes("OCRscore"))
}
\arguments{
\item{gene_matrix}{A data frame of expression counts where rows are genes, columns are samples, and the gene symbols are available in a column defined as "Gene".}

\item{ocr.genes}{A character vector of the genes to use when calculating the score. May need to be altered due to changing gene labels.}
}
\value{
A data frame with a column of samples and a column of their OCR-scores.
}
\description{
The OCR-score is the mean of log2(count + 1) across genes associated with
mitochondrial oxygen demand and hypoxia.
}
