% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkGenes.R
\name{checkGenes}
\alias{checkGenes}
\title{Check presence of required genes}
\usage{
checkGenes(genes, score = NULL, expected.genes = NULL)
}
\arguments{
\item{genes}{A character vector containing the genes available in the data set the calculation is to be performed on.}

\item{score}{The name of the score whose dependent genes should be checked.}

\item{expected.genes}{A vector of the genes expected in the data. Can be provided instead of score.}
}
\value{
A character vector containing any genes required by the score and missing in the data set.
}
\description{
Genes may appear to be missing from the data set due to changing naming conventions. This function returns gene names that are required for the specified calculation
so that they can be addressed, possibly by adding updated gene names to the vector before passing to the score calculator.
}
