% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_entries.R
\name{gene_entries}
\alias{gene_entries}
\title{Gene Entries}
\usage{
gene_entries(query)
}
\arguments{
\item{query}{Name of gene symbol which contains group of genes}
}
\value{
signature group
}
\description{
Gene Entries
}
