% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_avg_Z_score.R
\name{calculateAvgZScore}
\alias{calculateAvgZScore}
\title{Calculate Average Z Score
Calculates the average z score for any given genes within a gene score}
\usage{
calculateAvgZScore(gene_matrix, genes)
}
\arguments{
\item{gene_matrix}{a data frame where columns are samples and rows are genes. Gene symbols are expected as a column labeled "Gene.Symbol".}

\item{gene_list}{a group of genes within a certain gene symbol}
}
\value{
A data set with all the genes within the requested gene symbol
group and the average z score for each gene. Check if all the
genes within "gene_list" are present in "gene_matrix".
Function will return true if all genes are present and
return false if there are missing genes and list out missing genes.
}
\description{
Calculate Average Z Score
Calculates the average z score for any given genes within a gene score
}
\examples{
calculate_avg_z_score(gene_matrix = expressions, gene_list = c("RNF43","BMP4","TSPAN8","PPP1R1B","SLC44A4","C9orf152","VWA2","AXIN2","SP5","NKD1","CFTR"))
}
