% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneEntries.R
\name{geneEntries}
\alias{geneEntries}
\title{Gene Entries
A function that can be used to call of specific gene signatures. This functions is originally intended to be used with the calculateAvgZScore function}
\usage{
geneEntries(query)
}
\arguments{
\item{query}{Name of the signature group. If query is passed "all" function will return all the signature groups}
}
\value{
The genes within a desired signature group
}
\description{
Gene Entries
A function that can be used to call of specific gene signatures. This functions is originally intended to be used with the calculateAvgZScore function
}
\examples{
geneEntries("IFNg_18")
}
