% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_entries.R
\name{geneEntries}
\alias{geneEntries}
\title{Gene Entries}
\usage{
geneEntries(query)
}
\arguments{
\item{query}{Name of the signature group. If query is passed "all" function will return all the signature groups}
}
\value{
The genes within a desired signature group
}
\description{
Gene Entries
}
