% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputGenes.R
\name{inputGenes}
\alias{inputGenes}
\title{Pull gene inputs for various score calculations}
\usage{
inputGenes(query)
}
\arguments{
\item{query}{The name of the score whose associated genes are to be pulled. c("Mitoscore", "Buffa")}
}
\value{
A character vector of the genes symbols associated with the queried score
}
\description{
Pull gene inputs for various score calculations
}
